--- src/main/java/com/embabel/hub/tripper/config/SecurityConfig.java
@@ -1,20 +1,0 @@
-package com.embabel.hub.tripper.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-public class SecurityConfig {
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-            .authorizeHttpRequests(auth -> auth
-                .requestMatchers("/", "/login**", "/css/**", "/js/**", "/images/**", "/webjars/**").permitAll()
-                .anyRequest().authenticated()
-            )
-            .oauth2Login();
-        return http.build();
-    }
-}
