+++ src/main/kotlin/com/embabel/agent/web/security/SecurityConfig.kt
@@ -1,17 +1,38 @@
-package com.embabel.tripper.config
+package com.embabel.agent.web.security
 
+import com.embabel.common.util.loggerFor
+import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.web.SecurityFilterChain
 
+@ConfigurationProperties(prefix = "embabel.security")
+data class SecurityProperties(
+    val enabled: Boolean = true,
+)
+
 @Configuration
 @EnableWebSecurity
-class SecurityConfig(private val customOAuth2UserService: CustomOAuth2UserService) {
+class SecurityConfig(
+    private val customOAuth2UserService: CustomOAuth2UserService,
+    private val securityProperties: SecurityProperties,
+) {
+
+    private fun unsecuredFilterChain(http: HttpSecurity): SecurityFilterChain {
+        http
+            .authorizeHttpRequests { it.anyRequest().permitAll() }
+            .csrf { it.disable() }
+        return http.build()
+    }
 
     @Bean
     fun securityFilterChain(http: HttpSecurity): SecurityFilterChain {
+        if (!securityProperties.enabled) {
+            loggerFor<SecurityConfig>().warn("Security is disabled in configuration. This is not recommended for production environments.")
+            return unsecuredFilterChain(http)
+        }
         http
             .csrf { it.disable() } // For simplicity in demo applications
             .authorizeHttpRequests { authorize ->
@@ -32,7 +53,7 @@ class SecurityConfig(private val customOAuth2UserService: CustomOAuth2UserServic
                     .logoutSuccessUrl("/login?logout")
                     .permitAll()
             }
-        
+
         return http.build()
     }
 }
