+++ src/main/kotlin/com/embabel/agent/web/security/UserController.kt
@@ -1,4 +1,4 @@
-package com.embabel.tripper.web
+package com.embabel.agent.web.security
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.security.oauth2.core.user.OAuth2User
@@ -13,6 +13,9 @@ class UserController {
     fun getUserInfo(@AuthenticationPrincipal oAuth2User: OAuth2User, model: Model): String {
         model.addAttribute("name", oAuth2User.getAttribute<String>("name"))
         model.addAttribute("email", oAuth2User.getAttribute<String>("email"))
+        oAuth2User.attributes["picture"]?.let { picture ->
+            model.addAttribute("picture", picture)
+        }
         model.addAttribute("attributes", oAuth2User.attributes)
         return "common/user-info"
     }
