+++ src/main/resources/application.properties
@@ -12,6 +12,10 @@ spring.thymeleaf.check-template-location=true
 spring.web.resources.cache.period=0
 spring.web.resources.chain.cache=false
 
+
+# Set to true to enable security
+embabel.security.enabled=false
+
 # Google OAuth2 Configuration
 # To set up Google OAuth2:
 # 1. Go to https://console.cloud.google.com/
@@ -20,7 +24,6 @@ spring.web.resources.chain.cache=false
 # 4. Create an OAuth client ID (Web application type)
 # 5. Add authorized redirect URIs: http://localhost:8080/login/oauth2/code/google
 # 6. Copy the client ID and client secret below
-
 spring.security.oauth2.client.registration.google.client-id=${GOOGLE_CLIENT_ID}
 spring.security.oauth2.client.registration.google.client-secret=${GOOGLE_CLIENT_SECRET}
 spring.security.oauth2.client.registration.google.scope=email,profile
