+++ src/main/resources/templates/common/fragments/user.html
@@ -7,26 +7,94 @@
 </head>
 <body>
 <!-- User Authentication Fragment -->
-<div th:fragment="userAuth" style="display: flex; align-items: center; gap: 1rem;" sec:authorize="isAuthenticated()">
-        <span style="display: flex; align-items: center; gap: 0.5rem; color: #666;">
-            <a href="/user"><svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor">
+<div th:fragment="userAuth"
+     style="display: flex;
+            align-items: center;
+            gap: 1rem;
+            padding: 12px 18px;
+            background: rgba(15, 23, 42, 0.6);
+            backdrop-filter: blur(5px);
+            border: 1px solid rgba(148, 163, 184, 0.2);
+            border-radius: 12px;
+            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
+            transition: all 0.3s ease;
+            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+            line-height: 1.5;"
+     sec:authorize="isAuthenticated()"
+     onmouseover="this.style.transform='translateY(-2px)'; this.style.boxShadow='0 6px 20px rgba(0, 0, 0, 0.3)'; this.style.borderColor='rgba(96, 165, 250, 0.3)'"
+     onmouseout="this.style.transform='translateY(0)'; this.style.boxShadow='0 4px 12px rgba(0, 0, 0, 0.2)'; this.style.borderColor='rgba(148, 163, 184, 0.2)'">
+
+    <!-- User Info Section -->
+    <span style="display: flex;
+                 align-items: center;
+                 gap: 8px;
+                 color: #f1f5f9;
+                 font-weight: 600;
+                 font-size: 0.95rem;
+                 letter-spacing: 0.4px;">
+        <a href="/user"
+           style="display: flex;
+                  align-items: center;
+                  justify-content: center;
+                  width: 36px;
+                  height: 36px;
+                  background: linear-gradient(135deg, #486346, #5d7c56);
+                  border-radius: 50%;
+                  color: #fff;
+                  text-decoration: none;
+                  transition: all 0.3s ease;
+                  box-shadow: 0 2px 8px rgba(72, 99, 70, 0.3);"
+           title="View Profile"
+           onmouseover="this.style.transform='scale(1.1) translateY(-2px)'; this.style.boxShadow='0 4px 15px rgba(72, 99, 70, 0.4)'"
+           onmouseout="this.style.transform='scale(1) translateY(0)'; this.style.boxShadow='0 2px 8px rgba(72, 99, 70, 0.3)'">
+            <svg width="18" height="18" viewBox="0 0 24 24" fill="currentColor">
                 <path d="M12 12c2.21 0 4-1.79 4-4s-1.79-4-4-4-4 1.79-4 4 1.79 4 4 4zm0 2c-2.67 0-8 1.34-8 4v2h16v-2c0-2.66-5.33-4-8-4z"/>
             </svg>
-            </a>
-            <span sec:authentication="name">User</span>
-        </span>
+        </a>
+        <span sec:authentication="name"
+              style="color: #cbd5e1;
+                     font-size: 0.9rem;
+                     font-weight: 500;">User</span>
+    </span>
 
-    <form th:action="@{/logout}" method="post" style="margin: 0;">
+    <!-- Logout Form -->
+    <form th:action="@{/logout}"
+          method="post"
+          style="margin: 0; display: flex;">
         <button type="submit"
-                style="background: none; border: none; color: #666; cursor: pointer; padding: 0; display: flex; align-items: center;"
+                style="display: flex;
+                       align-items: center;
+                       justify-content: center;
+                       width: 36px;
+                       height: 36px;
+                       background: rgba(148, 163, 184, 0.1);
+                       border: 1px solid rgba(148, 163, 184, 0.2);
+                       border-radius: 10px;
+                       color: #94a3b8;
+                       cursor: pointer;
+                       transition: all 0.3s ease;
+                       backdrop-filter: blur(5px);
+                       font-family: inherit;
+                       padding: 0;"
                 title="Logout"
-                onmouseover="this.style.color='#333'"
-                onmouseout="this.style.color='#666'">
-            <svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor">
+                onmouseover="this.style.background='rgba(239, 68, 68, 0.15)';
+                           this.style.borderColor='rgba(239, 68, 68, 0.3)';
+                           this.style.color='#ef4444';
+                           this.style.transform='translateY(-2px)';
+                           this.style.boxShadow='0 4px 12px rgba(239, 68, 68, 0.2)'"
+                onmouseout="this.style.background='rgba(148, 163, 184, 0.1)';
+                          this.style.borderColor='rgba(148, 163, 184, 0.2)';
+                          this.style.color='#94a3b8';
+                          this.style.transform='translateY(0)';
+                          this.style.boxShadow='none'"
+                onmousedown="this.style.transform='translateY(-1px)'"
+                onmouseup="this.style.transform='translateY(-2px)'">
+            <svg width="18" height="18" viewBox="0 0 24 24" fill="currentColor">
                 <path d="M17 7l-1.41 1.41L18.17 11H8v2h10.17l-2.58 2.59L17 17l5-5zM4 5h8V3H4c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h8v-2H4V5z"/>
             </svg>
         </button>
     </form>
 </div>
+
 </body>
 </html>
