+++ src/main/resources/templates/common/user-info.html
@@ -16,33 +16,39 @@
                 <div class="card-body p-4">
                     <h2 class="text-center mb-4">User Profile</h2>
 
+                    <div class="text-center mb-4" th:if="${picture != null and !#strings.isEmpty(picture)}">
+                        <img th:src="${picture}"
+                             th:alt="${name + ' profile picture'}"
+                             class="profile-image"
+                             onerror="this.style.display='none'">
+                    </div>
                     <div class="mb-3">
                         <h4>Welcome, <span th:text="${name}">User</span>!</h4>
                         <p class="text-muted" th:text="${email}">user@example.com</p>
                     </div>
 
-                    <div class="mb-3">
-                        <h5>Your Account Details</h5>
-                        <table class="table table-striped">
-                            <thead>
-                            <tr>
-                                <th>Attribute</th>
-                                <th>Value</th>
-                            </tr>
-                            </thead>
-                            <tbody>
-                            <tr th:each="attr : ${attributes}">
-                                <td th:text="${attr.key}">Attribute</td>
-                                <td th:text="${attr.value}">Value</td>
-                            </tr>
-                            </tbody>
-                        </table>
-                    </div>
+                    <!--                    <div class="mb-3">-->
+                    <!--                        <h5>Your Account Details</h5>-->
+                    <!--                        <table class="table table-striped">-->
+                    <!--                            <thead>-->
+                    <!--                            <tr>-->
+                    <!--                                <th>Attribute</th>-->
+                    <!--                                <th>Value</th>-->
+                    <!--                            </tr>-->
+                    <!--                            </thead>-->
+                    <!--                            <tbody>-->
+                    <!--                            <tr th:each="attr : ${attributes}">-->
+                    <!--                                <td th:text="${attr.key}">Attribute</td>-->
+                    <!--                                <td th:text="${attr.value}">Value</td>-->
+                    <!--                            </tr>-->
+                    <!--                            </tbody>-->
+                    <!--                        </table>-->
+                    <!--                    </div>-->
 
                     <div class="d-flex justify-content-between mt-4">
-                        <a href="/" class="btn btn-primary">Back to Home</a>
+                        <a href="/" class="friendly-map-btn">Back to Home</a>
                         <form th:action="@{/logout}" method="post">
-                            <button type="submit" class="btn btn-outline-danger">Logout</button>
+                            <button type="submit" class="submit-btn">Logout</button>
                         </form>
                     </div>
                 </div>
